pyproject.toml
[tool.poetry]
name = "python-monorepo"  # Spaces replaced with hyphens
version = "0.1.0"
description = "This is my monorepo for personal projects"
authors = ["Joey <josephrobertlopez@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "*", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pyinstaller = "^6.11.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
behave = "^1.2.6"
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
flake8 = "^7.2.0"
pytest-cov = "^6.0.0"
pygame = "^2.1.3"

[tool.poetry.group.pomodoro.dependencies]
pygame = "^2.1.3"

[tool.pytest]
python_paths = ["/home/joey/Documents/GitHub/python-libs"]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
